==> Performance_benchmarks/Bulk_gtex/2_Gtex_DTU.Rmd <==
---
title: "2_Gtex_DTU"
author: "Jeroen Gilis"
date: "02/11/2020"
output: html_document
---

**In order to run this script (2_Gtex_DTU.Rmd), the dataset GTEx_benchmark_datasets_count.Rdata (or, alternatively, GTEx_benchmark_datasets_scaledTPM.Rdata) is required.** This file can either be generated with the 1_Gtex_prepare.Rmd script or downloaded from Zenodo.  

Run the DTU analyses for all 7 methods on all (18) Gtex benchmark datasets. Note that for DEXSeq, DRIMSeq and NBSplice we only run the datasets with 5 samples in each group, as these methods do not scale to large datasets.
This code runs approximately 4 hours on a MacBook Pro 2018, processor; 2,3 GHz Quad-Core Intel Core i5, 16GB RAM. Most of the runtime is attributed to the DEXSeq, DRIMSeq and NBSplice analyses.
Most of the runtime is attributed to the DEXSeq, DRIMSeq and NBSplice analyses.

**If you do not want to run this script, its output can also be downloaded from Zenodo: GTEx_DTU_results_count.Rdata (or, alternatively, GTEx_DTU_results_scaledTPM.Rdata) **

```{r setup, include=FALSE, echo=FALSE}
require("knitr")
opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# Load DTU methods

Source file that allows running all seven DTU methods that were assessed in this paper: satuRn, DoubleExpSeq, limma diffsplice, edgeRdiffsplice, DEXSeq, DRIMseq and NBSplice.

```{r, message=FALSE, warning=FALSE}
### load libraries
library(doMC)
library(edgeR)
library(limma)
library(DEXSeq)
library(DRIMSeq)
library(DoubleExpSeq)
library(NBSplice)
#devtools::install_local("/Users/jg/Desktop/PhD/DTU_project/satuRn", force = TRUE, quiet = FALSE)
library(satuRn)

source(file="./Performance_benchmarks/DTU_methods.R")
```

# Set up parallel execution

Note that the optimal parameter setting for parallelizing the DTU analyses depends on your specific computing system.

```
nrCores <- 2

if(nrCores != 1) {
    doParallel <- TRUE
    doProgress <- 'none'
    registerDoMC(cores = nrCores)
} else {
    doParallel <- FALSE
    doProgress <- 'text'
}
```

# Run all analyses 

Run the DTU analyses for all 7 methods on all (18) Gtex benchmark datasets. Note that for DEXSeq, DRIMSeq and NBSplice we only run the datasets with 5 samples in each group, as these methods do not scale to large datasets.
This code runs approximately 4 hours on a MacBook Pro 2018, processor; 2,3 GHz Quad-Core Intel Core i5, 16GB RAM. Most of the runtime is attributed to the DEXSeq, DRIMSeq and NBSplice analyses.

```{r}
### Load the benchmark datasets, which are output from 1_Gtex_prepare.Rmd
#load(file="./Data/GTEx_benchmark_datasets_count.Rdata")
load(file="./Data/GTEx_benchmark_datasets_scaledTPM.Rdata")

gtexBenchmarkData <- c(gtexBenchmarkDataLenient,gtexBenchmarkDataStringent)

### Run DTU analyses on benchmark data
    
print("start satuRn")

tStart <- Sys.time()
suppressWarnings(gtexDtuBenchmark_satuRn <- plyr::llply(
    .data = gtexBenchmarkData,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- satuRn_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
))
difftime(Sys.time(), tStart)
    
print("start DoubleExpSeq")

tStart <- Sys.time()
gtexDtuBenchmark_DoubleExpSeq <- plyr::llply(
    .data = gtexBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- DoubleExpSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design,
            quiet=FALSE
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)
    
print("start limma diffsplice")

tStart <- Sys.time()
gtexDtuBenchmark_limmaDiffsplice <- plyr::llply(
    .data = gtexBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- limma_diffsplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)
    
print("start edgeR_diffsplice")

tStart <- Sys.time()
gtexDtuBenchmark_edgeRdiffsplice <- plyr::llply(
    .data = gtexBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- edgeR_diffsplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)
 





   
print("start DEXSeq")

tStart <- Sys.time()
gtexDtuBenchmark_DEXSeq <- plyr::llply(
    .data = gtexBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
            
        if(ncol(localData$data) > 40) {
            return(NULL) 
        } # do not run DEXSeq for datasets with many samples - too slow
          
        ### Perform DTU analysis
        localRes <- DEXSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)
    
print("start DRIMSeq")

tStart <- Sys.time()
gtexDtuBenchmark_DRIMSeq <- plyr::llply(
    .data = gtexBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
          
        if(ncol(localData$data) > 40) {
            return(NULL) 
        } # do not run DRIMSeq for datasets with many samples - too slow
          
        ### Perform DTU analysis
        localRes <- DRIMSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)
    
     
print("start NBSplice")

tStart <- Sys.time()
gtexDtuBenchmark_NBSplice <- plyr::llply(
    .data = gtexBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
          
        if(ncol(localData$data) > 40) {
              return(NULL) 
        } # do not run NBSPlice for datasets with many samples - too slow
          
        ### Perform DTU analysis
        localRes <- NBSplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design,
            quiet=FALSE
        )

        if(is.null(localRes)){
            return( list(dtuAnalysis = localRes))
        }

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

### add method name to list names for easy post-analysis
names(gtexDtuBenchmark_satuRn) <- paste0('satuRn_',names(gtexDtuBenchmark_satuRn))
names(gtexDtuBenchmark_DoubleExpSeq) <- paste0('DoubleExpSeq_', names(gtexDtuBenchmark_DoubleExpSeq))
names(gtexDtuBenchmark_limmaDiffsplice) <- paste0('limma_diffsplice_', names(gtexDtuBenchmark_limmaDiffsplice))
names(gtexDtuBenchmark_edgeRdiffsplice) <- paste0('edgeR_diffsplice_', names(gtexDtuBenchmark_edgeRdiffsplice))
names(gtexDtuBenchmark_DEXSeq) <- paste0('DEXSeq_', names(gtexDtuBenchmark_DEXSeq))
names(gtexDtuBenchmark_DRIMSeq) <- paste0('DRIMSeq_', names(gtexDtuBenchmark_DRIMSeq))
names(gtexDtuBenchmark_NBSplice) <- paste0('NBSplice_', names(gtexDtuBenchmark_NBSplice))

### Save result
# save(gtexDtuBenchmark_satuRn,
#     gtexDtuBenchmark_limmaDiffsplice,
#     gtexDtuBenchmark_edgeRdiffsplice,
#     gtexDtuBenchmark_DoubleExpSeq,
#     gtexDtuBenchmark_DEXSeq,
#     gtexDtuBenchmark_DRIMSeq,
#     gtexDtuBenchmark_NBSplice,
#     file="./Data/GTEx_DTU_results_count.Rdata")

save(gtexDtuBenchmark_satuRn,
    gtexDtuBenchmark_limmaDiffsplice,
    gtexDtuBenchmark_edgeRdiffsplice,
    gtexDtuBenchmark_DoubleExpSeq,
    gtexDtuBenchmark_DEXSeq,
    gtexDtuBenchmark_DRIMSeq,
    gtexDtuBenchmark_NBSplice,
    file="./Data/GTEx_DTU_results_scaledTPM.Rdata")
```
